pyproject.toml
[project]
name = "devin"
version = "0.1.0"
description = "Devin is your very own AI command centre"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "openmeteo-requests>=1.7.2",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "rich>=14.1.0",
    "ruff>=0.12.11",
    "uvicorn>=0.35.0",
]
